Moved components into folders
