+++ recipeApp/client/src/app/app.component.html
@@ -1,4 +1,3 @@
 
-what the heck is going on
 <app-header></app-header>
 <router-outlet></router-outlet>
