+++ recipeApp/client/src/app/app.module.ts
@@ -27,14 +27,14 @@ import { MatSortModule } from '@angular/material/sort';
 import { MatTableModule } from '@angular/material/table'
 
 import { AppComponent } from './app.component';
-import { LoginEditorComponent } from './login-editor.component';
+import { LoginEditorComponent } from './login-editor/login-editor.component';
 import { AppHeaderComponent } from './app-header/app-header.component';
-import { WelcomeComponent } from './welcome.component';
+import { WelcomeComponent } from './welcome/welcome.component';
 import { AppRoutingModule } from './shared/app-router';
-import { registerFormComponent } from './registerform.component';
+import { registerFormComponent } from './registerform/registerform.component';
 import { MenuNavigationComponent} from './menu-navigation/menu-navigation.component';
-import { HomeComponent } from './home.component';
-import { RecipesComponent } from './recipes.component';
+import { HomeComponent } from './home/home.component';
+import { RecipesComponent } from './recipes/recipes.component';
 import { StepperComponent } from './stepper/stepper.component';
 import { MealPlanComponent } from './meal-plan/meal-plan.component';
 import { MealPlanPageComponent } from './meal-plan-page/meal-plan-page.component';
