+++ recipeApp/client/src/app/home/home.component.ts
@@ -1,12 +1,12 @@
 import {Component, OnInit} from '@angular/core'
 import { HttpClient } from '@angular/common/http';
-import {AuthService} from './shared/auth/auth.service';
+import {AuthService} from '../shared/auth/auth.service';
 import { Router } from '@angular/router';
 
 @Component({
     selector: 'home',
     templateUrl: './home.component.html',
-    styleUrls: ['./app.component.css']
+    styleUrls: ['../app.component.css']
 
 })
