+++ recipeApp/client/src/app/login-editor/login-editor.component.html
@@ -1,4 +1,3 @@
-<!-- <app-header [isUserLoggedIn]="false"></app-header> -->
 
 <mat-card><mat-card-content>Login to Mallow</mat-card-content></mat-card>
