+++ recipeApp/client/src/app/login-editor/login-editor.component.ts
@@ -2,13 +2,13 @@ import {Component, OnInit} from '@angular/core'
 import { FormGroup,FormControl,FormBuilder } from '@angular/forms'
 import { HttpClient} from '@angular/common/http';
 import { Router } from '@angular/router';
-import { AuthService } from './shared/auth/auth.service';
+import { AuthService } from '../shared/auth/auth.service';
 
 
 @Component({
     selector:'app-login-editor',
     templateUrl: './login-editor.component.html',
-    styleUrls: ['./app.component.css']
+    styleUrls: ['../app.component.css']
 })
 export class LoginEditorComponent implements OnInit {
     loginForm!: FormGroup
