+++ recipeApp/client/src/app/recipes/recipes.component.ts
@@ -2,7 +2,7 @@ import {Component, ViewChild, OnInit} from '@angular/core'
 import { NgForOf } from '@angular/common'
 import { HttpClient, HttpParams } from '@angular/common/http'
 import { MatPaginator, PageEvent } from '@angular/material/paginator'
-import {AuthService} from './shared/auth/auth.service'
+import {AuthService} from '../shared/auth/auth.service'
 
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Observable } from 'rxjs';
@@ -28,7 +28,7 @@ interface rCount {
 @Component({
     selector: 'app-recipes',
     templateUrl: './recipes.component.html',
-    styleUrls: ['./app.component.css']
+    styleUrls: ['../app.component.css']
 
 })
 
@@ -80,8 +80,6 @@ async loadItems() {
     
     this.backendItems =await this.httpClient.get<IRecipeItem[]>(URL,{params: params}).toPromise()
     this.httpClient.get<rCount>(`/server/recipecount`,{params:params}).subscribe((data)=>{this.totalRows = data.total})
-
-    
 
 }
