--- recipeApp/client/src/app/test.component.spec.ts
@@ -1,11 +1,0 @@
-/*import { TestComponent } from './test.component';
-
-describe('TestComponent', () => {
-
-    it('should be fine (this is to confirm the test is set up properly)', () => {
-        expect(true).toBeTruthy();
-    })
-
-    // INSERT UNIT TESTS AND DESCRIPTIONS BELOW
-
-})*/
