+++ recipeApp/client/src/app/welcome/welcome.component.ts
@@ -3,7 +3,7 @@ import {Component, OnInit, Input} from '@angular/core'
 @Component({
     selector: 'welcome',
     templateUrl: './welcome.component.html',
-    styleUrls: ['./app.component.css']
+    styleUrls: ['../app.component.css']
 
 })
